requirements.txt
opencv-python
tf-keras
deepface
pillow
pypng
pyqrcode
